refactor: 整句学习后端按职责拆分并修正命名
- completeSentence 与 isLearningAvailable 不依赖会话，拆出到新建的 SentenceLearningService；「云端 AI 未配置」文案上移到 AiProviderService 供两侧共用
- ChatSessionService.start 改名 sendMessage：会话早已由 create 建立，start 名不副实；IPC 路由 chat/start 同步改为 chat/send-message，类型改为 ChatSendMessageParams/Result，前端调用点一并更新
- ChatSessionStopParams 注释把「暂停」纠正为「取消」，避免误导有续跑语义
- 删除 isCancellation 纯转发包装与 runAnalysisStream 的 streamLogger 别名；startAnalysis 实现侧 JSDoc 与接口重复的部分收敛到接口
