+++ src/backend/ioc/inversify.config.ts
@@ -46,6 +46,7 @@ import FfmpegService, { FfmpegServiceImpl } from '@/backend/services/FfmpegServi
 import DpTaskService, { DpTaskServiceImpl } from '@/backend/services/DpTaskService';
 import ChatService, { ChatServiceImpl } from '@/backend/services/ChatService';
 import ChatSessionService, { ChatSessionServiceImpl } from '@/backend/services/ChatSessionService';
+import SentenceLearningService, { SentenceLearningServiceImpl } from '@/backend/services/SentenceLearningService';
 import ChatSessionStore, { InMemoryChatSessionStore } from '@/backend/services/chat/ChatSessionStore';
 import AiProviderService, { AiProviderServiceImpl } from '@/backend/services/AiProviderService';
 import ModelRoutingService, { ModelRoutingServiceImpl } from '@/backend/services/ModelRoutingService';
@@ -239,6 +240,7 @@ container.bind<DpTaskService>(TYPES.DpTaskService).to(DpTaskServiceImpl).inSingl
 container.bind<ChatService>(TYPES.ChatService).to(ChatServiceImpl).inSingletonScope();
 container.bind<ChatSessionStore>(TYPES.ChatSessionStore).to(InMemoryChatSessionStore).inSingletonScope();
 container.bind<ChatSessionService>(TYPES.ChatSessionService).to(ChatSessionServiceImpl).inSingletonScope();
+container.bind<SentenceLearningService>(TYPES.SentenceLearningService).to(SentenceLearningServiceImpl).inSingletonScope();
 container.bind<PlaybackRepairService>(TYPES.PlaybackRepairService).to(PlaybackRepairServiceImpl).inSingletonScope();
 container.bind<PlaybackCapabilityService>(TYPES.PlaybackCapabilityService).to(PlaybackCapabilityServiceImpl).inSingletonScope();
 container.bind<SplitVideoService>(TYPES.SplitVideoService).to(SplitVideoServiceImpl).inSingletonScope();
